MM/deployment/vps/requirements-vps.txt
numpy>=1.26.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
py-clob-client>=0.20.0
py-clob-client-v2>=0.1.0
playwright>=1.44.0
pytest>=8.0.0
